+++ examples/gsd/_03_gsd_evaluation.py
@@ -167,7 +167,7 @@ matches
 # often want to run a full evaluation on an entire dataset.
 # This can be done using the :class:`~mobgap.gsd.evaluation.GsdEvaluationPipeline` class and some ``tpcp`` functions.
 #
-# But lets start with selecting some data.
+# But let's start with selecting some data.
 # We want to use all the simulated real-world walking data from the INDIP reference system (Test11).
 simulated_real_world_walking = LabExampleDataset(reference_system="INDIP").get_subset(test="Test11")
 
@@ -176,16 +176,16 @@ simulated_real_world_walking
 # Now we can use the :class:`~mobgap.gsd.evaluation.GsdEvaluationPipeline` class to directly run a Gsd algorithm on
 # a datapoint.
 # The pipeline takes care of extracting the required data.
-from mobgap.gsd.evaluation import GsdEvaluationPipeline
+from mobgap.gsd.pipeline import GsdEmulationPipeline
 
-pipeline = GsdEvaluationPipeline(GsdIluz())
+pipeline = GsdEmulationPipeline(GsdIluz())
 
-pipeline.run(simulated_real_world_walking[0]).gs_list_
+pipeline.safe_run(simulated_real_world_walking[0]).gs_list_
 
 # %%
 # Note, that this did just "run" the pipeline on a single datapoint.
 # If we want to run it on all datapoints and evaluate the performance of the algorithm, we can use the
-# ``tpcp.validate.validate`` function.
+# :func:`~tpcp.validate.validate` function.
 #
 # It uses the build in ``score`` method of the pipeline to calculate the performance of the algorithm on each datapoint
 # and then takes the mean of the results.
@@ -205,7 +205,7 @@ evaluation_results["single_reference"][0][0]
 evaluation_results["single_detected"][0][0]
 
 # %%
-# If you want to calculate additional metrics, you can either create a custom score function or sublcass the pipeline
+# If you want to calculate additional metrics, you can either create a custom score function or subclass the pipeline
 # and overwrite the score function.
 #
 # Parameter Optimization
@@ -233,7 +233,7 @@ para_grid = ParameterGrid({"algo__window_length_s": [2, 3, 4]})
 
 cross_validate_results = pd.DataFrame(
     cross_validate(
-        GridSearch(GsdEvaluationPipeline(GsdIluz()), para_grid, return_optimized="precision"),
+        GridSearch(GsdEmulationPipeline(GsdIluz()), para_grid, return_optimized="precision"),
         simulated_real_world_walking,
         cv=3,
         return_train_score=True,