+++ mobgap/utils/array_handling.py
@@ -61,15 +61,16 @@ def sliding_window_view(data: np.ndarray, window_size_samples: int, overlap_samp
     return view
 
 
-def _loc_with_empty_fallback(df: pd.DataFrame, name: Any) -> pd.DataFrame:
+def _get_group_with_empty_fallback(
+    group: pd.core.groupby.DataFrameGroupBy, original_df: pd.DataFrame, name: Any
+) -> pd.DataFrame:
     try:
-        return df.loc[name]
+        return group.get_group(name)
     except KeyError:
         # We return a frame that has the same columns as the original, but no rows.
         # We also replicate the index, just without any rows.
-
-        index = df.index[:0]
-        return pd.DataFrame(columns=df.columns, index=index)
+        index = original_df.index[:0]
+        return pd.DataFrame(columns=original_df.columns, index=index)
 
 
 class MultiGroupBy:
@@ -83,6 +84,7 @@ class MultiGroupBy:
     """
 
     _primary_groupby: pd.core.groupby.DataFrameGroupBy
+    _secondary_groupbys: list[pd.core.groupby.DataFrameGroupBy]
 
     def __init__(
         self,
@@ -121,8 +123,20 @@ class MultiGroupBy:
             self._primary_groupby = self.primary_df.groupby(level=self.groupby)
         return self._primary_groupby
 
+    @property
+    def secondary_groupbys(self) -> list[pd.core.groupby.DataFrameGroupBy]:
+        """The secondary groupby objects.
+
+        These are the groupers created from the secondary dataframes.
+        """
+        if not hasattr(self, "_secondary_groupbys"):
+            self._secondary_groupbys = [df.groupby(level=self.groupby) for df in self.secondary_dfs]
+        return self._secondary_groupbys
+
     def _get_secondary_vals(self, name: Union[str, tuple[str, ...]]) -> list[pd.DataFrame]:
-        return [_loc_with_empty_fallback(df, [name]) for df in self.secondary_dfs]
+        return [
+            _get_group_with_empty_fallback(g, df, name) for g, df in zip(self.secondary_groupbys, self.secondary_dfs)
+        ]
 
     def get_group(self, name: Union[str, tuple[str, ...]]) -> tuple[pd.DataFrame, ...]:
         """Get an individual group by name.
@@ -131,8 +145,7 @@ class MultiGroupBy:
         -------
         A tuple containing the groups from each dataframe.
         """
-        return_val = self.primary_groupby.get_group(name), *self._get_secondary_vals(name)
-        return return_val
+        return self.primary_groupby.get_group(name), *self._get_secondary_vals(name)
 
     @property
     def groups(
@@ -154,7 +167,7 @@ class MultiGroupBy:
         self,
     ) -> Iterator[tuple[Union[str, tuple[str, ...]], tuple[pd.DataFrame, ...]]]:
         """Iterate over the groups and return a tuple with the group name and the group dataframes."""
-        return ((name, self.get_group(name)) for name, group in self.primary_groupby)
+        return ((name, self.get_group(name)) for name, _ in self.primary_groupby)
 
     def apply(self, func: Callable, *args: Unpack[list[Any]], **kwargs: Unpack[dict[str, Any]]) -> pd.DataFrame:
         """Apply a function that takes the group values from each df as input.